fedimint_core
Module admin_client

Admin (guardian) client types

Modules

serde_tls_cert 🔒

Structs

ConfigGenConnectionsRequest

Sent by admin user to the API

ConfigGenParamsConsensus

The config gen params that need to be in consensus, sent by the config gen leader to all the other guardians

ConfigGenParamsRequest

Config gen params that can be configured from the UI

ConfigGenParamsResponse

The config gen params response which includes our peer id

PeerServerParams

Connection information sent between peers in order to start config gen

Enums

ServerStatus

The state of the server returned via APIs