fedimint_core::config
Trait ISupportedDkgMessage

pub trait ISupportedDkgMessage:
    Sized
    + Serialize
    + DeserializeOwned {
    // Required methods
    fn to_msg(self) -> SupportedDkgMessage;
    fn from_msg(msg: SupportedDkgMessage) -> Result<Self>;
}

Supported (by Fedimint’s code) DkgMessage<T> types

Since DkgMessage is an open-set, yet we only use a subset of it, we can make a subset-trait to convert it to an enum that we it’s easier to handle.

Candidate for refactoring after modularization effort is complete.

Required Methods

fn to_msg(self) -> SupportedDkgMessage

fn from_msg(msg: SupportedDkgMessage) -> Result<Self>

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.