Type Alias fedimint_core::config::CommonModuleInitRegistry

pub type CommonModuleInitRegistry = ModuleInitRegistry<DynCommonModuleInit>;

Aliased Type

struct CommonModuleInitRegistry(BTreeMap<ModuleKind, DynCommonModuleInit, Global>);

Fields
0: BTreeMap<ModuleKind, DynCommonModuleInit, Global>
Implementations

impl<M> ModuleInitRegistry<M>

pub fn new() -> Self

pub fn attach<T>(&mut self, gen: T)where T: Into<M> + 'static + Send + Sync, M: AsRef<dyn IDynCommonModuleInit + Send + Sync + 'static>,

pub fn kinds(&self) -> BTreeSet<ModuleKind>

pub fn get(&self, k: &ModuleKind) -> Option<&M>

impl<M> ModuleInitRegistry<M>where M: AsRef<dyn IDynCommonModuleInit + Send + Sync + 'static>,

pub fn decoders<'a>( &self, modules: impl Iterator<Item = (ModuleInstanceId, &'a ModuleKind)> ) -> Result<ModuleDecoderRegistry>

๐Deprecated: You probably want available_decoders to support missing module kinds. If you really want a strict behavior, use decoders_strict

pub fn decoders_strict<'a>( &self, modules: impl Iterator<Item = (ModuleInstanceId, &'a ModuleKind)> ) -> Result<ModuleDecoderRegistry>

Get decoders for modules and fail if any is unsupported

pub fn available_decoders<'a>( &self, modules: impl Iterator<Item = (ModuleInstanceId, &'a ModuleKind)> ) -> Result<ModuleDecoderRegistry>

Get decoders for modules and skip unsupported ones