Type Definition fedimint_core::config::ServerModuleGenParamsRegistry

pub type ServerModuleGenParamsRegistry = ModuleGenRegistry<ConfigGenModuleParams>;

Configs for each module’s DKG

Note: in the future, we should make this one a ModuleRegistry<ConfigGenParams>, as each module instance will need a distinct config for dkg.