fedimint_core
Module core

Fundamental types Fedimint Core API (common) module interface

Fedimint supports externally implemented modules.

This (Rust) module defines common interoperability types and functionality that is used on both client and sever side.

Modules

backup

Structs

Decoder

Consensus encoding decoder for module-specific types

DecoderBuilder

DynClientConfig

An owned, immutable input to a Transaction

DynInput

An owned, immutable input to a Transaction

DynInputError

DynModuleConsensusItem

An owned, immutable output of a Transaction before it was finalized

DynOutput

An owned, immutable output of a Transaction

DynOutputError

DynOutputOutcome

An owned, immutable output of a Transaction before it was finalized

DynUnknown

A type used by when decoding dyn-types, when the module is missing

ModuleKind

A type of a module

OperationId

Unique identifier for one semantic, correlatable operation.

OperationIdFullFmt

OperationIdShortFmt

Enums

FinalizationError

Constants

LEGACY_HARDCODED_INSTANCE_ID_MINT

LEGACY_HARDCODED_INSTANCE_ID_WALLET

MODULE_INSTANCE_ID_GLOBAL

Special IDs we use for global dkg

Traits

ClientConfig

IClientConfig

IInput

Something that can be an DynInput in a Transaction

IInputError

IModuleConsensusItem

IOutput

Something that can be an DynOutput in a Transaction

IOutputError

IOutputOutcome

Input

InputError

IntoDynInstance

A type that has a Dyn*, type erased version of itself

ModuleConsensusItem

Output

OutputError

OutputOutcome

Type Aliases

DecodeFn ๐

ModuleInstanceId

Module instance ID