Trait fedimint_core::core::IntoDynInstance

pub trait IntoDynInstance {
    type DynType: 'static;

    // Required method
    fn into_dyn(self, instance_id: ModuleInstanceId) -> Self::DynType;
}

A type that has a Dyn*, type erased version of itself

Use IntoDynNever in places where a given type will never actually be created, but something is needed to appease the type system.

Required Associated Types

type DynType: 'static

The type erased version of the type implementing this trait

Required Methods

fn into_dyn(self, instance_id: ModuleInstanceId) -> Self::DynType

Convert self into its type-erased equivalent