fedimint_core::db
Function remove_current_db_version_if_exists

async fn remove_current_db_version_if_exists(
    version_dbtx: &mut DatabaseTransaction<'_>,
    is_new_db: bool,
    target_db_version: DatabaseVersion,
) -> DatabaseVersion

Removes DatabaseVersion from DatabaseVersionKeyV0 if it exists and returns the current database version. If the current version does not exist, use target_db_version if the database is new. Otherwise, return DatabaseVersion(0) to ensure all migrations are run.