fedimint_core::db

Trait AutocommitResultExt

pub trait AutocommitResultExt<T, E> {
    // Required method
    fn unwrap_autocommit(self) -> Result<T, E>;
}

Required Methods

fn unwrap_autocommit(self) -> Result<T, E>

Unwraps the “commit failed” error variant. Use this in cases where autocommit is instructed to run indefinitely and commit will thus never fail.