fedimint_core::db

Trait IDatabaseTransactionOps

pub trait IDatabaseTransactionOps: IDatabaseTransactionOpsCore + MaybeSend { }

Additional operations (only some) database transactions expose, on top of IDatabaseTransactionOpsCore

In certain contexts exposing these operations would be a problem, so they are moved to a separate trait.