fedimint_core

Module envs

Common environment variables

Structs

BitcoinRpcConfig

Configuration for the bitcoin RPC

Constants

FM_BITCOIND_COOKIE_FILE_ENV

Env var that can be set to point at the bitcoind’s cookie file to use for auth

FM_BITCOIN_POLLING_INTERVAL_SECS_ENV

Env var how often to poll bitcoin source

FM_BITCOIN_RPC_KIND_ENV

Env var for bitcoin RPC kind (obsolete, use FM_DEFAULT_* instead)

FM_BITCOIN_RPC_URL_ENV

Env var for bitcoin URL (obsolete, use FM_DEFAULT_* instead)

FM_DEFAULT_BITCOIN_RPC_KIND_BAD_ENV

FM_DEFAULT_BITCOIN_RPC_KIND_ENV

Env var for bitcoin RPC kind (default, used only as a default value for DKG config settings)

FM_DEFAULT_BITCOIN_RPC_URL_BAD_ENV

FM_DEFAULT_BITCOIN_RPC_URL_ENV

Env var for bitcoin URL (default, used only as a default value for DKG config settings)

FM_ENABLE_MODULE_LNV2_ENV

FM_FORCE_BITCOIN_RPC_KIND_BAD_ENV

FM_FORCE_BITCOIN_RPC_KIND_ENV

Env var for bitcoin RPC kind (forced, takes priority over config settings)

FM_FORCE_BITCOIN_RPC_URL_BAD_ENV

FM_FORCE_BITCOIN_RPC_URL_ENV

Env var for bitcoin URL (default, takes priority over config settings)

FM_IN_DEVIMINT_ENV

devimint will set when code is running inside devimint

FM_IROH_API_SECRET_KEY_OVERRIDE_ENV

FM_IROH_CONNECT_OVERRIDES_ENV

Env var to override iroh connectivity

FM_IROH_P2P_SECRET_KEY_OVERRIDE_ENV

FM_USE_UNKNOWN_MODULE_ENV

In tests we want to routinely enable an extra unknown module to ensure all client code handles correct modules that client doesn’t know about.

FM_WALLET_FEERATE_SOURCES_ENV

List of json api endpoint sources to use as a source of fee rate estimation.

FM_WS_API_CONNECT_OVERRIDES_ENV

Env var to override tcp api connectivity

Functions

is_env_var_set

Check if env variable is set and not equal 0 or false which are common ways to disable something.

is_rbf_withdrawal_enabled

Use to allow process_output to process RBF withdrawal outputs.

is_running_in_test_env

Use to detect if running in a test environment, either cargo test or devimint.

parse_kv_list_from_env