Module fedimint_core::fmt_utils

Formatting helpers

Structs

AbbreviateDebug

A wrapper that causes the inner T to be debug-formatted using AbbreviatedDebug

AbbreviateHexBytes

Use for displaying bytes in the logs

AbbreviateJson

Use for displaying potentially large [serde_json::Value]s in the logs

OptStacktrace

Optional stacktrace formatting for errors.

Traits

AbbreviatedDebug

Something that can be debug-formatted in an abbreviated way

Functions

floor_char_boundary

fmt_abbreviated_object 🔒

Format json object value truncating keys if there’s too many, and keys and values if they are too long

fmt_abbreviated_str 🔒

Format json string value if it’s too long

fmt_abbreviated_vec 🔒

Format json array value truncating elements if there’s too many, and values if they are too long

rust_log_full_enabled