fedimint_core

Module iroh_prod

We want to provide our own infrastructure to be independent of n0, but for now we don’t use the discovery part of it pending figuring out some performance regressions.

Constants

FM_IROH_DNS_FEDIMINT_PROD

The Iroh/Pkarr DNS server hosted by Fedimint project

FM_IROH_RELAYS_FEDIMINT_PROD

The Iroh relays hosted by Fedimint project