Trait fedimint_core::module::ModuleGen

pub trait ModuleGen: Debug + Sized {
    type Decoder: Decoder;

    const KIND: ModuleKind;
    const DATABASE_VERSION: DatabaseVersion;

    // Required methods
    fn decoder(&self) -> Self::Decoder;
    fn versions(&self, core: CoreConsensusVersion) -> &[ModuleConsensusVersion];
    fn init<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        cfg: ServerModuleConfig,
        db: Database,
        env: &'life1 BTreeMap<OsString, OsString>,
        task_group: &'life2 mut TaskGroup
    ) -> Pin<Box<dyn Future<Output = Result<DynServerModule>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn trusted_dealer_gen(
        &self,
        peers: &[PeerId],
        params: &ConfigGenParams
    ) -> BTreeMap<PeerId, ServerModuleConfig>;
    fn distributed_gen<'life0, 'life1, 'life2, 'life3, 'life4, 'life5, 'async_trait>(
        &'life0 self,
        connections: &'life1 MuxPeerConnections<ModuleInstanceId, DkgPeerMsg>,
        our_id: &'life2 PeerId,
        module_id: ModuleInstanceId,
        peers: &'life3 [PeerId],
        params: &'life4 ConfigGenParams,
        task_group: &'life5 mut TaskGroup
    ) -> Pin<Box<dyn Future<Output = Result<Cancellable<ServerModuleConfig>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             'life3: 'async_trait,
             'life4: 'async_trait,
             'life5: 'async_trait;
    fn to_config_response(&self, config: Value) -> Result<ModuleConfigResponse>;
    fn validate_config(
        &self,
        identity: &PeerId,
        config: ServerModuleConfig
    ) -> Result<()>;
    fn hash_client_module(&self, config: Value) -> Result<Hash>;
    fn dump_database<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        dbtx: &'life1 mut DatabaseTransaction<'life2>,
        prefix_names: Vec<String>
    ) -> Pin<Box<dyn Future<Output = Box<dyn Iterator<Item = (String, Box<dyn Serialize + Send>)> + '_>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;

    // Provided method
    fn get_database_migrations(&self) -> MigrationMap<'_> { ... }
}

Module Generation trait with associated types

Needs to be implemented by module generation type

For examples, take a look at one of the MintConfigGenerator, WalletConfigGenerator, or LightningConfigGenerator structs.

Required Associated Types

type Decoder: Decoder

Required Associated Constants

const KIND: ModuleKind

const DATABASE_VERSION: DatabaseVersion

This represents the module’s database version that the current code is compatible with. It is important to increment this value whenever a key or a value that is persisted to the database within the module changes. It is also important to add the corresponding migration function in get_database_migrations which should define how to move from the previous database version to the current version.

Required Methods

fn decoder(&self) -> Self::Decoder

fn versions(&self, core: CoreConsensusVersion) -> &[ModuleConsensusVersion]

Version of the module consensus supported by this implementation given a certain CoreConsensusVersion.

Refer to ModuleConsensusVersion for more information about versioning.

One module implementation (ModuleGen of a given ModuleKind) can potentially implement multiple versions of the consensus, and depending on the config module instance config, instantiate the desired one. This method should expose all the available versions, purely for information, setup UI and sanity checking purposes.

fn init<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, cfg: ServerModuleConfig, db: Database, env: &'life1 BTreeMap<OsString, OsString>, task_group: &'life2 mut TaskGroup ) -> Pin<Box<dyn Future<Output = Result<DynServerModule>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Initialize the DynServerModule instance from its config

fn trusted_dealer_gen( &self, peers: &[PeerId], params: &ConfigGenParams ) -> BTreeMap<PeerId, ServerModuleConfig>

fn distributed_gen<'life0, 'life1, 'life2, 'life3, 'life4, 'life5, 'async_trait>( &'life0 self, connections: &'life1 MuxPeerConnections<ModuleInstanceId, DkgPeerMsg>, our_id: &'life2 PeerId, module_id: ModuleInstanceId, peers: &'life3 [PeerId], params: &'life4 ConfigGenParams, task_group: &'life5 mut TaskGroup ) -> Pin<Box<dyn Future<Output = Result<Cancellable<ServerModuleConfig>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, 'life3: 'async_trait, 'life4: 'async_trait, 'life5: 'async_trait,

fn to_config_response(&self, config: Value) -> Result<ModuleConfigResponse>

fn validate_config( &self, identity: &PeerId, config: ServerModuleConfig ) -> Result<()>

fn hash_client_module(&self, config: Value) -> Result<Hash>

fn dump_database<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, dbtx: &'life1 mut DatabaseTransaction<'life2>, prefix_names: Vec<String> ) -> Pin<Box<dyn Future<Output = Box<dyn Iterator<Item = (String, Box<dyn Serialize + Send>)> + '_>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Provided Methods

fn get_database_migrations(&self) -> MigrationMap<'_>

Retreives the MigrationMap from the module to be applied to the database before the module is initialized. The MigrationMap is indexed on the from version.