Trait fedimint_core::module::ServerModuleGen

pub trait ServerModuleGen: ExtendsCommonModuleGen + Sized {
    const DATABASE_VERSION: DatabaseVersion;

    fn versions(&self, core: CoreConsensusVersion) -> &[ModuleConsensusVersion];
    fn init<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        cfg: ServerModuleConfig,
        db: Database,
        env: &'life1 BTreeMap<OsString, OsString>,
        task_group: &'life2 mut TaskGroup
    ) -> Pin<Box<dyn Future<Output = Result<DynServerModule>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait;
    fn trusted_dealer_gen(
        &self,
        peers: &[PeerId],
        params: &ConfigGenModuleParams
    ) -> BTreeMap<PeerId, ServerModuleConfig>;
    fn distributed_gen<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        peer: &'life1 PeerHandle<'_>,
        params: &'life2 ConfigGenModuleParams
    ) -> Pin<Box<dyn Future<Output = DkgResult<ServerModuleConfig>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait;
    fn to_config_response(&self, config: Value) -> Result<ModuleConfigResponse>;
    fn validate_config(
        &self,
        identity: &PeerId,
        config: ServerModuleConfig
    ) -> Result<()>;
    fn dump_database<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        dbtx: &'life1 mut ModuleDatabaseTransaction<'life2>,
        prefix_names: Vec<String>
    ) -> Pin<Box<dyn Future<Output = Box<dyn Iterator<Item = (String, Box<dyn Serialize + Send>)> + '_>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait;

    fn kind() -> ModuleKind { ... }
    fn get_database_migrations(&self) -> MigrationMap<'_> { ... }
}

Module Generation trait with associated types

Needs to be implemented by module generation type

For examples, take a look at one of the MintConfigGenerator, WalletConfigGenerator, or LightningConfigGenerator structs.

Required Associated Constants

const DATABASE_VERSION: DatabaseVersion

This represents the module’s database version that the current code is compatible with. It is important to increment this value whenever a key or a value that is persisted to the database within the module changes. It is also important to add the corresponding migration function in get_database_migrations which should define how to move from the previous database version to the current version.

Required Methods

fn versions(&self, core: CoreConsensusVersion) -> &[ModuleConsensusVersion]

Version of the module consensus supported by this implementation given a certain CoreConsensusVersion.

Refer to ModuleConsensusVersion for more information about versioning.

One module implementation (ServerModuleGen of a given ModuleKind) can potentially implement multiple versions of the consensus, and depending on the config module instance config, instantiate the desired one. This method should expose all the available versions, purely for information, setup UI and sanity checking purposes.

fn init<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
cfg: ServerModuleConfig,
db: Database,
env: &'life1 BTreeMap<OsString, OsString>,
task_group: &'life2 mut TaskGroup
) -> Pin<Box<dyn Future<Output = Result<DynServerModule>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,

Initialize the DynServerModule instance from its config

fn trusted_dealer_gen(
&self,
peers: &[PeerId],
params: &ConfigGenModuleParams
) -> BTreeMap<PeerId, ServerModuleConfig>

fn distributed_gen<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
peer: &'life1 PeerHandle<'_>,
params: &'life2 ConfigGenModuleParams
) -> Pin<Box<dyn Future<Output = DkgResult<ServerModuleConfig>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,

fn to_config_response(&self, config: Value) -> Result<ModuleConfigResponse>

fn validate_config(
&self,
identity: &PeerId,
config: ServerModuleConfig
) -> Result<()>

fn dump_database<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
dbtx: &'life1 mut ModuleDatabaseTransaction<'life2>,
prefix_names: Vec<String>
) -> Pin<Box<dyn Future<Output = Box<dyn Iterator<Item = (String, Box<dyn Serialize + Send>)> + '_>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,

Provided Methods

fn kind() -> ModuleKind

fn get_database_migrations(&self) -> MigrationMap<'_>

Retrieves the MigrationMap from the module to be applied to the database before the module is initialized. The MigrationMap is indexed on the from version.