fedimint_core::net

Module peers

Modules

fake

Enums

Recipient

This enum defines the intended recipient of a p2p message.

Traits

IP2PConnections

Connection manager that tries to keep connections open to all peers

Type Aliases

DynP2PConnections