Module fedimint_core::session_outcome

Atomic BFT unit containing consensus items

Structs

AcceptedItem

If two correct nodes obtain two ordered items from the broadcast they are guaranteed to be in the same order. However, an ordered items is only guaranteed to be seen by all correct nodes if a correct node decides to accept it.

SchnorrSignature

SessionOutcome

All items ordered in a session that have been accepted by Fedimint Consensus are recorded in the corresponding block. A running Federation produces a SessionOutcome roughly every five minutes. Therefore, just like in Bitcoin, a SessionOutcome might be empty if no items are ordered in that time or all ordered items are discarded by Fedimint Consensus.

SignedSessionOutcome

A signed block combines a block with the naive threshold secp schnorr signature for its header created by the federation. The signed blocks allow clients and recovering guardians to verify the federations consensus history. After a signed block has been created it is stored in the database.

Enums

SessionStatus