Trait fedimint_core::api::GlobalFederationApi

pub trait GlobalFederationApi {
    fn submit_transaction<'life0, 'async_trait>(
        &'life0 self,
        tx: Transaction
    ) -> Pin<Box<dyn Future<Output = FederationResult<TransactionId>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn fetch_tx_outcome<'life0, 'life1, 'async_trait>(
        &'life0 self,
        txid: &'life1 TransactionId
    ) -> Pin<Box<dyn Future<Output = FederationResult<Option<TransactionStatus>>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait;
    fn await_tx_outcome<'life0, 'life1, 'async_trait>(
        &'life0 self,
        txid: &'life1 TransactionId
    ) -> Pin<Box<dyn Future<Output = FederationResult<TransactionStatus>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait;
    fn fetch_epoch_history<'life0, 'life1, 'async_trait>(
        &'life0 self,
        epoch: u64,
        epoch_pk: PublicKey,
        decoders: &'life1 ModuleDecoderRegistry
    ) -> Pin<Box<dyn Future<Output = FederationResult<SignedEpochOutcome>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait;
    fn fetch_epoch_count<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = FederationResult<u64>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn fetch_output_outcome<'life0, 'life1, 'async_trait, R>(
        &'life0 self,
        out_point: OutPoint,
        module_decoder: &'life1 Decoder
    ) -> Pin<Box<dyn Future<Output = Result<Option<R>, OutputOutcomeError>> + Send + 'async_trait>>
    where
        R: OutputOutcome + 'async_trait,
        Self: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait;
    fn await_output_outcome<'life0, 'life1, 'async_trait, R>(
        &'life0 self,
        outpoint: OutPoint,
        timeout: Duration,
        module_decoder: &'life1 Decoder
    ) -> Pin<Box<dyn Future<Output = Result<R, OutputOutcomeError>> + Send + 'async_trait>>
    where
        R: OutputOutcome + 'async_trait,
        Self: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait;
    fn download_client_config<'life0, 'life1, 'async_trait>(
        &'life0 self,
        info: &'life1 WsClientConnectInfo
    ) -> Pin<Box<dyn Future<Output = FederationResult<ClientConfig>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait;
    fn consensus_config_hash<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = FederationResult<Hash>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait;
    fn upload_backup<'life0, 'life1, 'async_trait>(
        &'life0 self,
        request: &'life1 SignedBackupRequest
    ) -> Pin<Box<dyn Future<Output = FederationResult<()>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait;
    fn download_backup<'life0, 'life1, 'async_trait>(
        &'life0 self,
        id: &'life1 XOnlyPublicKey
    ) -> Pin<Box<dyn Future<Output = FederationResult<Vec<ClientBackupSnapshot>>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait;
}

The API for the global (non-module) endpoints

Required Methods

fn submit_transaction<'life0, 'async_trait>(
&'life0 self,
tx: Transaction
) -> Pin<Box<dyn Future<Output = FederationResult<TransactionId>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

fn fetch_tx_outcome<'life0, 'life1, 'async_trait>(
&'life0 self,
txid: &'life1 TransactionId
) -> Pin<Box<dyn Future<Output = FederationResult<Option<TransactionStatus>>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,

fn await_tx_outcome<'life0, 'life1, 'async_trait>(
&'life0 self,
txid: &'life1 TransactionId
) -> Pin<Box<dyn Future<Output = FederationResult<TransactionStatus>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,

fn fetch_epoch_history<'life0, 'life1, 'async_trait>(
&'life0 self,
epoch: u64,
epoch_pk: PublicKey,
decoders: &'life1 ModuleDecoderRegistry
) -> Pin<Box<dyn Future<Output = FederationResult<SignedEpochOutcome>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,

fn fetch_epoch_count<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = FederationResult<u64>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

fn fetch_output_outcome<'life0, 'life1, 'async_trait, R>(
&'life0 self,
out_point: OutPoint,
module_decoder: &'life1 Decoder
) -> Pin<Box<dyn Future<Output = Result<Option<R>, OutputOutcomeError>> + Send + 'async_trait>>where
R: OutputOutcome + 'async_trait,
Self: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,

fn await_output_outcome<'life0, 'life1, 'async_trait, R>(
&'life0 self,
outpoint: OutPoint,
timeout: Duration,
module_decoder: &'life1 Decoder
) -> Pin<Box<dyn Future<Output = Result<R, OutputOutcomeError>> + Send + 'async_trait>>where
R: OutputOutcome + 'async_trait,
Self: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,

fn download_client_config<'life0, 'life1, 'async_trait>(
&'life0 self,
info: &'life1 WsClientConnectInfo
) -> Pin<Box<dyn Future<Output = FederationResult<ClientConfig>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,

Fetch client configuration info only if verified against a federation id

fn consensus_config_hash<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = FederationResult<Hash>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,

Fetches the server consensus hash if enough peers agree on it

fn upload_backup<'life0, 'life1, 'async_trait>(
&'life0 self,
request: &'life1 SignedBackupRequest
) -> Pin<Box<dyn Future<Output = FederationResult<()>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,

fn download_backup<'life0, 'life1, 'async_trait>(
&'life0 self,
id: &'life1 XOnlyPublicKey
) -> Pin<Box<dyn Future<Output = FederationResult<Vec<ClientBackupSnapshot>>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,