fedimint_core

Module config

Federation configuration

Modules

serde_binary_human_readable

Structs

ClientConfig

Total client config

ClientConfigV0

Total client config v0 (<0.4.0). Does not contain broadcast public keys.

ClientModuleConfig

Config for the client-side of a particular Federation module

ConfigGenModuleParams

Type erased ModuleInitParams used to generate the ServerModuleConfig during config gen

EmptyGenParams

Empty struct for if there are no params

FederationId

The federation id is a copy of the authentication threshold public key of the federation

FederationIdPrefix

Prefix of the FederationId, useful for UX improvements

GlobalClientConfig

Federation-wide client config

GlobalClientConfigV0

Federation-wide client config

JsonClientConfig

Client config that cannot be cryptographically verified but is easier to parse by external tools

JsonWithKind

serde_json::Value that must contain kind: String field

ModuleInitRegistry

PeerUrl

ServerModuleConfig

Config for the server-side of a particular Federation module

ServerModuleConsensusConfig

Enums

DkgMessage

P2PMessage

Constants

ALEPH_BFT_UNIT_BYTE_LIMIT

This limits the RAM consumption of a AlephBFT Unit to roughly 50kB

META_FEDERATION_NAME_KEY

Key under which the federation name can be sent to client in the meta part of the config

Traits

ModuleInitParams

TypedServerModuleConfig

Module (server side) config, typed

TypedServerModuleConsensusConfig

Consensus-critical part of a server side module config

Functions

de_int_key 🔒

load_from_file

optional_de_int_key 🔒

Type Aliases

CommonModuleInitRegistry

ServerModuleConfigGenParamsRegistry

Registry that contains the config gen params for all modules