Module fedimint_core::config

Federation configuration

Modules

serde_binary_human_readable

serde_commit 🔒

Handling the Group serialization with a wrapper

Structs

ClientConfig

Total client config

ClientModuleConfig

Config for the client-side of a particular Federation module

ConfigGenModuleParams

Type erased ModuleInitParams used to generate the ServerModuleConfig during config gen

EmptyGenParams

Empty struct for if there are no params

FederationId

The federation id is a copy of the authentication threshold public key of the federation

FederationIdPrefix

Prefix of the FederationId, useful for UX improvements

GlobalClientConfig

Federation-wide client config

JsonClientConfig

Client config that cannot be cryptographically verified but is easier to parse by external tools

JsonWithKind

serde_json::Value that must contain kind: String field

ModuleInitRegistry

PeerUrl

ServerModuleConfig

Config for the server-side of a particular Federation module

ServerModuleConsensusConfig

Enums

DkgError

Captures an error occurring in DKG

DkgMessage

DkgPeerMsg

Things that a distributed_gen config can send between peers

SupportedDkgMessage

enum version of SupportedDkgMessage

Constants

ALEPH_BFT_UNIT_BYTE_LIMIT

This limits the RAM consumption of a AlephBFT Unit to roughly 50kB

META_FEDERATION_NAME_KEY

Key under which the federation name can be sent to client in the meta part of the config

META_OVERRIDE_URL_KEY

Key under which the override URL can be sent to client in the meta part of the config

META_VETTED_GATEWAYS_KEY

Key under which the vetted gateways can be sent to client in the meta part of the config

Traits

DkgGroup

Defines a group (e.g. G1 or G2) that we can generate keys for

ISupportedDkgMessage

Supported (by Fedimint’s code) DkgMessage<T> types

ModuleInitParams

SGroup

TypedServerModuleConfig

Module (server side) config, typed

TypedServerModuleConsensusConfig

Consensus-critical part of a server side module config

Functions

de_int_key 🔒

load_from_file

Type Aliases

CommonModuleInitRegistry

DkgResult

Result of running DKG

ServerModuleConfigGenParamsRegistry

Registry that contains the config gen params for all modules

ServerModuleInitRegistry