Trait fedimint_core::config::ISupportedDkgMessage

pub trait ISupportedDkgMessage:
    Sized
    + Serialize
    + DeserializeOwned {
    // Required methods
    fn to_msg(self) -> SupportedDkgMessage;
    fn from_msg(msg: SupportedDkgMessage) -> Result<Self>;
}

Supported (by Fedimint’s code) DkgMessage<T> types

Since DkgMessage is an open-set, yet we only use a subset of it, we can make a subset-trait to convert it to an enum that we it’s easier to handle.

Candidate for refactoring after modularization effort is complete.

Required Methods

fn to_msg(self) -> SupportedDkgMessage

fn from_msg(msg: SupportedDkgMessage) -> Result<Self>

Object Safety

This trait is not object safe.