Module fedimint_core::core

Fedimint Core API (common) module interface

Fedimint supports externally implemented modules.

This (Rust) module defines common interoperability types and functionality that is used on both client and sever side.

Modules

backup

client

server

Fedimint Core Server module interface

Structs

Decoder

Decoder for module associated types

DecoderBuilder

DynClientConfig

An owned, immutable input to a Transaction

DynInput

An owned, immutable input to a Transaction

DynModuleConsensusItem

An owned, immutable output of a Transaction before it was finalized

DynOutput

An owned, immutable output of a Transaction

DynOutputOutcome

An owned, immutable output of a Transaction before it was finalized

KeyPair

Opaque data structure that holds a keypair consisting of a secret and a public key.

ModuleKind

A type of a module

Enums

FinalizationError

Constants

LEGACY_HARDCODED_INSTANCE_ID_LN

LEGACY_HARDCODED_INSTANCE_ID_MINT

LEGACY_HARDCODED_INSTANCE_ID_WALLET

MODULE_INSTANCE_ID_GLOBAL

Special IDs we use for global dkg

Traits

ClientConfig

IClientConfig

IInput

Something that can be an DynInput in a Transaction

IModuleConsensusItem

IOutput

Something that can be an DynOutput in a Transaction

IOutputOutcome

Input

IntoDynInstance

A type that has a Dyn*, type erased version of itself

ModuleConsensusItem

Output

OutputOutcome

Type Definitions

ModuleInstanceId

Module instance ID