Trait fedimint_core::db::ISingleUseDatabaseTransaction

pub trait ISingleUseDatabaseTransaction<'a>: 'a + MaybeSend + IDatabaseTransactionOps<'a> {
    // Required methods
    fn commit_tx<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn add_notification_key(&mut self, _key: &[u8]) -> Result<()>;
}

ISingleUseDatabaseTransaction re-defines the functions from IDatabaseTransaction but does not consumed self when committing to the database. This allows for wrapper structs to more easily borrow ISingleUseDatabaseTransaction without needing to make additional allocations.

Required Methods

fn commit_tx<'life0, 'async_trait>( &'life0 mut self ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn add_notification_key(&mut self, _key: &[u8]) -> Result<()>