Module fedimint_core::encoding

This module defines a binary encoding interface which is more suitable for consensus critical encoding than e.g. bincode. Over time all structs that need to be encoded to binary will be migrated to this interface.

Modules

as_hex

Serde implementations using hex-encoded encodables

btc 🔒

secp256k1 🔒

tbs 🔒

tls 🔒

Macros

impl_encode_decode_num_as_bigsize 🔒

impl_encode_decode_num_as_plain 🔒

impl_encode_decode_tuple 🔒

Structs

BigSize

Lightning TLV uses a custom variable-length integer called BigSize. It is similar to Bitcoin’s variable-length integers except that it is serialized in big-endian instead of little-endian.

CountWrite

A writer counting number of writes written to it

DecodeError

Enums

DynRawFallback

A type that decodes module_instance_id-prefixed Ts even when corresponding Decoder is not available.

Traits

Decodable

Data which can be encoded in a consensus-consistent way

DynEncodable

Object-safe trait for things that can encode themselves

Encodable

Data which can be encoded in a consensus-consistent way

Derive Macros

Decodable

Encodable

UnzipConsensus