fedimint_core::encoding

Module btc

Macros

impl_encode_decode_bridge 🔒

Structs

BitoinIoWriteAdapter

A writer counting number of bytes written to it

BufBitcoinReader 🔒

Wrap buffering support for implementations of Read. A reader which keeps an internal buffer to avoid hitting the underlying stream directly for every read.

NetworkLegacyEncodingWrapper

Wrapper around bitcoin::Network that encodes and decodes the network as a little-endian u32. This is here for backwards compatibility and is used by the LNv1 and WalletV1 modules.

SimpleBitcoinRead 🔒