fedimint_core::encoding

Trait Encodable

pub trait Encodable {
    // Required method
    fn consensus_encode<W: Write>(&self, writer: &mut W) -> Result<(), Error>;

    // Provided methods
    fn consensus_encode_to_vec(&self) -> Vec<u8>  { ... }
    fn consensus_encode_to_hex(&self) -> String { ... }
    fn consensus_encode_to_len(&self) -> u64 { ... }
    fn consensus_hash<H>(&self) -> H
       where H: Hash,
             H::Engine: Write { ... }
    fn consensus_hash_sha256(&self) -> Hash { ... }
}

Data which can be encoded in a consensus-consistent way

Required Methods

fn consensus_encode<W: Write>(&self, writer: &mut W) -> Result<(), Error>

Encode an object with a well-defined format. Returns the number of bytes written on success.

The only errors returned are errors propagated from the writer.

Provided Methods

fn consensus_encode_to_vec(&self) -> Vec<u8>

Self::consensus_encode to newly allocated Vec<u8>

fn consensus_encode_to_hex(&self) -> String

Encode and convert to hex string representation

fn consensus_encode_to_len(&self) -> u64

Encode without storing the encoding, return the size

fn consensus_hash<H>(&self) -> H
where H: Hash, H::Engine: Write,

Generate a SHA256 hash of the consensus encoding using the default hash engine for H.

Can be used to validate all federation members agree on state without revealing the object

fn consensus_hash_sha256(&self) -> Hash

Self::consensus_hash for [bitcoin::hashes::sha256::Hash]

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.