Trait fedimint_core::encoding::Encodable

pub trait Encodable {
    // Required method
    fn consensus_encode<W: Write>(&self, writer: &mut W) -> Result<usize, Error>;

    // Provided methods
    fn consensus_encode_to_vec(&self) -> Result<Vec<u8>, Error> { ... }
    fn consensus_encode_to_hex(&self) -> Result<String, Error> { ... }
    fn consensus_hash<H>(&self) -> H
       where H: Hash,
             H::Engine: Write { ... }
}

Data which can be encoded in a consensus-consistent way

Required Methods

fn consensus_encode<W: Write>(&self, writer: &mut W) -> Result<usize, Error>

Encode an object with a well-defined format. Returns the number of bytes written on success.

The only errors returned are errors propagated from the writer.

Provided Methods

fn consensus_encode_to_vec(&self) -> Result<Vec<u8>, Error>

Self::consensus_encode to newly allocated Vec<u8>

fn consensus_encode_to_hex(&self) -> Result<String, Error>

fn consensus_hash<H>(&self) -> Hwhere H: Hash, H::Engine: Write,

Generate a SHA256 hash of the consensus encoding using the default hash engine for H.

Can be used to validate all federation members agree on state without revealing the object