Module fedimint_core::invite_code

Federation invite code

Structs

InviteCode

Information required for client to join Federation

Enums

InviteCodeData 🔒

Data that can be encoded in the invite code. Currently we always just use one Api and one FederationId variant in an invite code, but more can be added in the future while still keeping the invite code readable for older clients, which will just ignore the new fields.

Constants

BECH32_HRP 🔒

We can represent client invite code as a bech32 string for compactness and error-checking