Module fedimint_core::module

Re-exports

pub use self::version::*;

Modules

audit

registry

version

Fedimint consensus and API versioning.

Macros

api_endpoint

Example

Structs

ApiAuth

Authentication uses the hashed user password in PHC format

ApiEndpoint

Definition of an API endpoint defined by a module M.

ApiEndpointContext

State made available to all API endpoints for handling a request

ApiError

ApiRequest

All requests from client to server contain these fields

DynCommonModuleInit

DynServerModuleInit

InputMeta

PeerHandle

A handle passed to ServerModuleInit::distributed_gen

SerdeModuleEncoding

Creates a struct that can be used to make our module-decodable structs interact with serde-based APIs (HBBFT, jsonrpsee). It creates a wrapper that holds the data as serialized

ServerModuleInitArgs

TransactionItemAmount

Information about the amount represented by an input or output.

Enums

ConsensusProposal

ModuleError

Traits

CommonModuleInit

ExtendsCommonModuleInit

IDynCommonModuleInit

Operations common to Server and Client side module gen dyn newtypes

IServerModuleInit

Interface for Module Generation

IntoModuleError

Extension trait with a function to map Results used by modules to ModuleError

ModuleCommon

Module associated types required by both client and server

ServerModule

ServerModuleInit

Module Generation trait with associated types

TypedApiEndpoint

Type Aliases

ApiRequestErased

HandlerFn 🔒

HandlerFnReturn 🔒