Trait fedimint_core::module::IDynCommonModuleInit

pub trait IDynCommonModuleInit: Debug {
    // Required methods
    fn decoder(&self) -> Decoder;
    fn module_kind(&self) -> ModuleKind;
    fn to_dyn_common(&self) -> DynCommonModuleInit;
}

Operations common to Server and Client side module gen dyn newtypes

Due to conflict of impl Trait for T for both ServerModuleInit and ClientModuleInit, we can’t really have a ICommonModuleInit, so to unify them in ModuleInitRegistry we move the common functionality to be an interface over their dyn newtype wrappers. A bit weird, but works.

Required Methods

fn decoder(&self) -> Decoder

fn module_kind(&self) -> ModuleKind

fn to_dyn_common(&self) -> DynCommonModuleInit