Module fedimint_core::runtime

Runtime (wasm32 vs native) differences handling Copyright 2021 The Matrix.org Foundation C.I.C. Abstraction over an executor so we can spawn tasks under WASM the same way we do usually.

Modules

impl 🔒

Structs

Elapsed

JoinError

Task failed to execute to completion.

JoinHandle

An owned permission to join on a task (await its termination).

Functions

block_in_place

block_on

sleep

sleep_until

spawn

timeout