Module fedimint_core::util

General purpose utilities

Modules

backoff_util

broadcaststream

Copied from tokio_stream 0.1.12 to use our optional Send bounds

update_merge

Structs

SafeUrl

Wrapper for Url that only prints the scheme, domain, port and path portion of a Url in its Display implementation. This is useful to hide private information like user names and passwords in logs or UIs.

Spanned

Enums

SafeUrlError

Traits

NextOrPending

Functions

handle_version_hash_command

For CLIs, detects version-hash as a single argument, prints the provided version hash, then exits the process.

retry

Run the supplied closure op_fn until it succeeds. Frequency and number of retries is determined by the specified strategy.

write_new

Write out a new file (like std::fs::write but fails if file already exists)

write_new_async

write_overwrite

write_overwrite_async

Type Aliases

BoxFuture

Future that is Send unless targeting WASM

BoxStream

Stream that is Send unless targeting WASM