fedimint_core::util

Module backoff_util

Traits

Backoff

Backoff is an Iterator that returns Duration.

Functions

aggressive_backoff

A backoff strategy for relatively quick foreground operations.

aggressive_backoff_long

api_networking_backoff

background_backoff

Backoff strategy for background tasks.

custom_backoff

fibonacci_max_one_hour

Retry every max 10s for up to one hour, with a more aggressive fibonacci backoff in the beginning to reduce expected latency.